fix: honor layer bounds for knockout when transparency doesn't shape
- Use the pixel layer's full bounds as the knockout shape when
  "transparency shapes layer" is off, instead of punching the whole tile
- Enable test support on recipe collection chips, cards, and remove buttons
- Render the recipe capture view above the library header
- Update tests for BlendMode's move to emulsion_raster and gpui_kit::px
- Collapse nested if-let blocks into let chains

+++ crates/emulsion-ui/src/editor/styles_ui.rs
@@ -611,7 +611,7 @@ impl EditorView {
                 body = body.child(label(title, p));
                 let points = [range.black, range.black_fade, range.white_fade, range.white];
                 let mut gradient = div()
-                    .id(("blend-if-gradient", backdrop))
+                    .id(("blend-if-gradient", usize::from(backdrop)))
                     .relative()
                     .w_full()
                     .h(px(30.))
